Move extractLogs method
